+++ simvue/config/__init__.py
@@ -5,5 +5,3 @@ Simvue Configuration
 This module contains definitions for the Simvue configuration options
 
 """
-
-from .user import SimvueConfiguration as SimvueConfiguration