+++ simvue/utilities.py
@@ -1,8 +1,8 @@
-import configparser
 import datetime
 import hashlib
 import logging
 import json
+import sys
 import tabulate
 import pydantic
 import importlib.util
@@ -15,6 +15,7 @@ import typing
 import jwt
 
 from datetime import timezone
+
 
 CHECKSUM_BLOCK_SIZE = 4096
 EXTRAS: tuple[str, ...] = ("plot", "torch")
@@ -137,10 +138,15 @@ def check_extra(extra_name: str) -> typing.Callable:
                 raise RuntimeError(
                     f"Plotting features require the '{extra_name}' extension to Simvue"
                 )
-            elif extra_name == "torch" and not importlib.util.find_spec("torch"):
-                raise RuntimeError(
-                    "PyTorch features require the 'torch' module to be installed"
-                )
+            elif extra_name == "torch":
+                if importlib.util.find_spec("torch"):
+                    raise RuntimeError(
+                        "PyTorch features require the 'torch' module to be installed"
+                    )
+                if sys.version_info.minor > 12:
+                    raise RuntimeError(
+                        "PyTorch features are not yet supported for python>3.12"
+                    )
             elif extra_name not in EXTRAS:
                 raise RuntimeError(f"Unrecognised extra '{extra_name}'")
             return class_func(self, *args, **kwargs) if class_func else None
@@ -254,30 +260,6 @@ def prettify_pydantic(class_func: typing.Callable) -> typing.Callable:
             raise RuntimeError(error_str)
 
     return wrapper
-
-
-def get_offline_directory() -> str:
-    """
-    Get directory for offline cache
-    """
-    directory = None
-
-    for filename in (
-        os.path.join(os.path.expanduser("~"), ".simvue.ini"),
-        "simvue.ini",
-    ):
-        if not filename or not os.path.exists(filename):
-            continue
-
-        with contextlib.suppress(Exception):
-            config = configparser.ConfigParser()
-            config.read(filename)
-            directory = config.get("offline", "cache")
-
-    if not (directory := os.environ.get("SIMVUE_OFFLINE_DIRECTORY", directory)):
-        directory = os.path.join(os.path.expanduser("~"), ".simvue")
-
-    return directory
 
 
 def create_file(filename: str) -> None: