+++ tests/conftest.py
@@ -59,9 +59,9 @@ def create_test_run(request) -> typing.Generator[typing.Tuple[sv_run.Run, dict],
 
 
 @pytest.fixture
-def create_test_run_offline(mocker: pytest_mock.MockerFixture, request) -> typing.Generator[typing.Tuple[sv_run.Run, dict], None, None]:
-    with tempfile.TemporaryDirectory() as offline_dir:
-        os.environ["SIMVUE_OFFLINE_DIRECTORY"] = offline_dir
+def create_test_run_offline(mocker: pytest_mock.MockerFixture, request, monkeypatch: pytest.MonkeyPatch) -> typing.Generator[typing.Tuple[sv_run.Run, dict], None, None]:
+    with tempfile.TemporaryDirectory() as temp_d:
+        monkeypatch.setenv("SIMVUE_OFFLINE_DIRECTORY", temp_d)
         with sv_run.Run("offline") as run:
             yield run, setup_test_run(run, True, request)
     clear_out_files()
@@ -82,9 +82,9 @@ def create_pending_run(request) -> typing.Generator[typing.Tuple[sv_run.Run, dic
 
 
 @pytest.fixture
-def create_plain_run_offline(mocker: pytest_mock.MockerFixture, request) -> typing.Generator[typing.Tuple[sv_run.Run, dict], None, None]:
-    with tempfile.TemporaryDirectory() as offline_dir:
-        os.environ["SIMVUE_OFFLINE_DIRECTORY"] = offline_dir
+def create_plain_run_offline(mocker: pytest_mock.MockerFixture, request, monkeypatch: pytest.MonkeyPatch) -> typing.Generator[typing.Tuple[sv_run.Run, dict], None, None]:
+    with tempfile.TemporaryDirectory() as temp_d:
+        monkeypatch.setenv("SIMVUE_OFFLINE_DIRECTORY", temp_d)
         with sv_run.Run("offline") as run:
             yield run, setup_test_run(run, False, request)
     clear_out_files()