+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+* Add support for defining Simvue run defaults using `tool.simvue` in a project `pyproject.toml` file.
+* Drop support for INI based configuration files.
+* Retrieve all metric values if `max_points` is unspecified or set to `None`.
+
+## [v1.1.1](https://github.com/simvue-io/client/releases/tag/v1.1.1) - 2024-10-22
+
+* Add missing `offline.cache` key to TOML config.
+* Fix repetition of server URL validation for each call to configuration.
+
+## [v1.1.0](https://github.com/simvue-io/client/releases/tag/v1.1.0) - 2024-10-21
+
 * Add option to specify a callback executed when an alert is triggered for a run.
 * Allow retrieval of all alerts when no constraints are specified.
 * Add carbon emissions statistics as optional metrics.